+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed [here](https://github.com/sensu-plugins/community/blob/master/HOW_WE_CHANGELOG.md)
 
 ## [Unreleased]
+- `check-puppet-errors.rb` will now also check for errors in the `last_run_report.yaml` (@elfranne)
 
 ## [3.1.0]
 ### Changed